Fe optimizations

Approved-by: Salman Shan

+++ src/api/request.js
@@ -1,16 +1,17 @@
 import axios from "axios";
-import { BASE_URL } from "../constants/api";
+import { isClient } from "../helpers/isClient";
 
 const request = axios.create({
-  baseURL: BASE_URL,
+  baseURL: process.env.NEXT_PUBLIC_BASE_URL,
   timeout: 30000,
 });
 
 request.interceptors.request.use(
   (config) => {
-    const user = JSON.parse(localStorage.getItem("user"));
-
-    config.headers["Authorization"] = `Bearer ${user?.access_token}`;
+    const user = JSON.parse(isClient ? localStorage.getItem("user") : "{}");
+    if (user?.access_token) {
+      config.headers["Authorization"] = `Bearer ${user?.access_token}`;
+    }
 
     if (config.url.startsWith("/sales/")) {
       const brand = JSON.parse(localStorage.getItem("brand"));
@@ -30,8 +31,8 @@ request.interceptors.response.use(
   },
   (error) => {
     if (error?.response?.status === 401) {
-      window.location.assign("/login");
-      localStorage.clear();
+      isClient && window.location.assign("/login");
+      isClient && localStorage.clear();
     }
     return error.response;
   }
